Fix filename case for DropboxUploadService.js
Apparently git will helpfully (?!) set core.ignorecase to 'true' when
you execute git init on a case insensitive filesystem [1].

So when the changes made for commit efd0078417fd renamed the file from
lowerCamelCase to CamelCase (as the exported item  was now a class) git
didn't record the change in filename case... causing issues importing
the CamerCase filename on case sensitive filesystems.

Hence, tests failed on build server (which is a case sensitive
filesystem).

Fix was to run an explicit 'git mv' command changing the case.

[1] https://git-scm.com/docs/git-config#git-config-coreignoreCase