Update Gandi DNS API in generate-ssl-cert.yml
It seems like the API has changed slightly. While the old API seems to
still be there, switching to the new API means we can do away with
having to remember the zone id... which I can't remember.

Probably good to stay up to date with the documented APIs too!