Use circleci python container rather than installing on base image
Should be quicker, and simpler to track python versions.

Since we no longer need to install python to be able to install AWS CLI,
it feels unnecessary to use the ci-install-awscli script - instead we
now just call pip install directly in the ci setup script (it needs to
be installed with --user as the global Python package location is not
owned by our ci user).