Fix SPA not loading Dropbox auth redirect URL
Redirecting back from Dropbox to the auth endpoint with token was failing
because the ListBucket permission did not exist so the attempt to load an
unknown resource gave a 403.

Browsers that had visited the site before seemed to still work as they
apparently used a cache from the service worker. Whereas new browsers (or
incognito) would give error.

Not sure how it ever worked. Either AWS changed something, or I got it to work
once then changed something and was relying on the cache since. Either way,
this seems like the right fix!